Develop

+++ backend/controller/user_controller.py
@@ -37,7 +37,9 @@ class UserController(MethodView):
             '/login': self.login,
             '/logout': self.logout,
             '/resetPassword': self.reset_password,
-            '/uploadFile': self.upload_user_file
+            '/uploadFile': self.upload_user_file,
+            '/archiveUser': self.archive_user,
+            '/unarchiveUser': self.unarchive_user
         }
         return self._dispatch_request(endpoint_mapping)
 
@@ -49,12 +51,14 @@ class UserController(MethodView):
         """
         endpoint_mapping = {
             '/getProfile': self.get_profile,
+            '/getContractInfo': self.get_contract_info,
             '/getUsers': self.get_users,
             '/getUsersByRole': self.get_users_by_role,
             '/getFile': self.get_user_file,
             '/getHiwis': self.get_hiwis,
             '/getSupervisor': self.get_supervisor,
-            '/getSupervisors': self.get_supervisors
+            '/getSupervisors': self.get_supervisors,
+            '/getArchivedUsers': self.get_archived_users
         }
         return self._dispatch_request(endpoint_mapping)
 
@@ -83,7 +87,6 @@ class UserController(MethodView):
             if request_path.endswith(path):
                 return func()
         return jsonify('Endpoint not found'), 404
-
     @jwt_required()
     @check_access(roles=[UserRole.ADMIN])
     def create_user(self):
@@ -92,6 +95,8 @@ class UserController(MethodView):
 
         :return: JSON response containing the status message and status code.
         """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
         if not request.is_json:
             return jsonify('Request data must be in JSON format'), 400
         user_data = request.get_json()
@@ -106,6 +111,8 @@ class UserController(MethodView):
 
         :return: JSON response containing the status message and status code.
         """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
         if not request.is_json:
             return jsonify('Request data must be in JSON format'), 400
         user_data = request.get_json()
@@ -121,10 +128,46 @@ class UserController(MethodView):
 
         :return: JSON response containing the status message and status code.
         """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
         if not request.is_json:
             return jsonify('Request data must be in JSON format'), 400
         username_data = request.get_json()
         result = self.user_service.delete_user(username_data['username'])
+        return jsonify(result.message), result.status_code
+
+    @jwt_required()
+    @check_access(roles=[UserRole.ADMIN])
+    def archive_user(self):
+        """
+        Archives a user identified by their username provided in JSON data.
+
+        :return: JSON response containing the status message and status code.
+        """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
+        if not request.is_json:
+            return jsonify('Request data must be in JSON format'), 400
+        data = request.get_json()
+        username = data.get('username')
+        result = self.user_service.archive_user(username)
+        return jsonify(result.message), result.status_code
+
+    @jwt_required()
+    @check_access(roles=[UserRole.ADMIN])
+    def unarchive_user(self):
+        """
+        Unarchives a user identified by their username provided in JSON data.
+
+        :return: JSON response containing the status message and status code.
+        """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
+        if not request.is_json:
+            return jsonify('Request data must be in JSON format'), 400
+        data = request.get_json()
+        username = data.get('username')
+        result = self.user_service.unarchive_user(username)
         return jsonify(result.message), result.status_code
 
     def login(self):
@@ -136,6 +179,8 @@ class UserController(MethodView):
         if not request.is_json:
             return jsonify('Request data must be in JSON format'), 400
         credentials = request.get_json()
+        if self.user_service.is_archived(credentials['username']):
+            return jsonify('User is archived'), 400
         result = self.auth_service.login(credentials['username'], credentials['password'])
         if not result.is_successful:
             return jsonify(result.message), result.status_code
@@ -162,6 +207,10 @@ class UserController(MethodView):
             return jsonify('Request data must be in JSON format'), 400
         credentials = request.get_json()
         username = credentials.get('username', get_jwt_identity())
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
+        if self.user_service.is_archived(username):
+            return jsonify('Password of archived user cannot be reset'), 400
         result = self.auth_service.reset_password(get_jwt_identity(), username, credentials['password'])
         return jsonify(result.message), result.status_code
 
@@ -172,8 +221,28 @@ class UserController(MethodView):
 
         :return: JSON response containing the user's profile data.
         """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
         user_profile = self.user_service.get_profile(get_jwt_identity())
         return jsonify(user_profile.to_dict()), 200
+
+    @jwt_required()
+    @check_access(roles=[UserRole.SUPERVISOR, UserRole.ADMIN, UserRole.SECRETARY])
+    def get_contract_info(self):
+        """
+        Retrieves the contract information for a given hiwi.
+        :return: JSON response containing the user's contract information.
+        """
+        args = request.args
+        if 'username' not in args:
+            return jsonify('Username parameter is required'), 400
+        username = args['username']
+        if not username:
+            return jsonify('Username is required'), 400
+        result = self.user_service.get_contract_info(username)
+        if not result.is_successful:
+            return jsonify(result.message), result.status_code
+        return jsonify(result.data.to_dict()), 200
 
     @jwt_required()
     @check_access(roles=[UserRole.ADMIN])
@@ -183,7 +252,23 @@ class UserController(MethodView):
 
         :return: JSON response containing a list of user profiles.
         """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
         users = self.user_service.get_users()
+        user_dict_list = [user.to_dict() for user in users]
+        return jsonify(user_dict_list), 200
+
+    @jwt_required()
+    @check_access(roles=[UserRole.ADMIN])
+    def get_archived_users(self):
+        """
+        Retrieves a list of all archived users in the system.
+
+        :return: JSON response containing a list of archived user profiles.
+        """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
+        users = self.user_service.get_archived_users()
         user_dict_list = [user.to_dict() for user in users]
         return jsonify(user_dict_list), 200
 
@@ -195,6 +280,8 @@ class UserController(MethodView):
 
         :return: JSON response containing a list of user profiles filtered by role.
         """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
         args = request.args
         if 'role' not in args:
             return jsonify('Role parameter is required'), 400
@@ -215,6 +302,9 @@ class UserController(MethodView):
         username = get_jwt_identity()
         file = request.files.get('file')
 
+        if self.user_service.is_archived(username):
+            return jsonify('User is archived'), 400
+
         file_type_str = request.form.get('fileType') or request.args.get('fileType')
 
         if not file or not file_type_str:
@@ -223,7 +313,6 @@ class UserController(MethodView):
         file_type = FileType.get_type_by_value(file_type_str)
         if not file_type:
             return jsonify("Invalid file type"), 400
-
         result = self.file_service.upload_image(file, username, file_type)
 
         return jsonify(result.message), result.status_code
@@ -236,22 +325,23 @@ class UserController(MethodView):
         :return: The file as a download if found, or a JSON response indicating an error with an
         appropriate HTTP status code if not found or if parameters are missing.
         """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
         username = request.args.get('username')
         file_type = FileType.get_type_by_value(request.args.get('fileType'))
 
+        if self.user_service.is_archived(username):
+            return jsonify('Cannot get file of an archived user'), 400
+
         if file_type == FileType.SIGNATURE and username != get_jwt_identity():
             return jsonify('You are not authorized to access this file'), 403
-
         if not username:
             return jsonify('Username is required'), 400
-
         if not file_type:
             return jsonify('Valid File type is required'), 400
-
         file_stream = self.file_service.get_image(username, file_type)
         if not file_stream:
             return jsonify('File not found'), 404
-
         return send_file(
             file_stream,
             as_attachment=True,
@@ -267,7 +357,11 @@ class UserController(MethodView):
         :return: A JSON response with the list of Hiwis and an appropriate HTTP status code.
         """
         username = get_jwt_identity()
+        if self.user_service.is_archived(username):
+            return jsonify('User is archived'), 400
         result = self.user_service.get_hiwis(username)
+        if not result.is_successful:
+            return jsonify(result.message), result.status_code
         hiwis_data = [hiwi.to_dict() for hiwi in result.data]
         return jsonify(hiwis_data), result.status_code
 
@@ -281,6 +375,8 @@ class UserController(MethodView):
         :return: A JSON response with the supervisor data and an appropriate HTTP status code.
         """
         username = get_jwt_identity()
+        if self.user_service.is_archived(username):
+            return jsonify('User is archived'), 400
         user = self.user_service.get_profile(username)
         request_args = request.args
 
@@ -305,6 +401,8 @@ class UserController(MethodView):
 
         :return: A JSON response with the list of Supervisors and an appropriate HTTP status code.
         """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
         result = self.user_service.get_supervisors()
         if not result.is_successful:
             return jsonify(result.message), result.status_code
@@ -320,7 +418,11 @@ class UserController(MethodView):
 
         :return: A JSON response with the result of the deletion attempt and an appropriate HTTP status code.
         """
+        if self.user_service.is_archived(get_jwt_identity()):
+            return jsonify('User is archived'), 400
         username = request.args.get('username')
+        if self.user_service.is_archived(username):
+            return jsonify('Cannot delete file of an archived user'), 400
         file_type = FileType.get_type_by_value(request.args.get('fileType'))
         if not username:
             return jsonify('Username is required'), 400
